+++ test/backgroundErrorTest.js
@@ -1,6 +1,6 @@
 const PgBoss = require('../')
 
-describe('maintenance error handling', function () {
+describe('background processing error handling', function () {
   this.retries(1)
 
   it('maintenance error handling works', function (done) {
@@ -13,17 +13,13 @@ describe('maintenance error handling', function () {
     const config = { ...this.test.bossConfig, ...defaults }
     const boss = new PgBoss(config)
 
-    const onError = (err) => {
-      if (err && boss.isStarted) {
-        boss.stop().then(() => done()).catch(done)
-      } else {
-        done(err)
-      }
-    }
+    boss.on('error', async () => {
+      boss.removeAllListeners()
+      await boss.stop()
+      done()
+    })
 
-    boss.on('error', onError)
-
-    boss.start().catch(done)
+    boss.start()
   })
 
   it('state monitoring error handling works', function (done) {
@@ -36,16 +32,25 @@ describe('maintenance error handling', function () {
     const config = { ...this.test.bossConfig, ...defaults }
     const boss = new PgBoss(config)
 
-    const onError = (err) => {
-      if (err && boss.isStarted) {
-        boss.stop().then(() => done()).catch(done)
-      } else {
-        done(err)
-      }
-    }
+    boss.on('error', async () => {
+      boss.removeAllListeners()
+      await boss.stop()
+      done()
+    })
 
-    boss.on('error', onError)
+    boss.start()
+  })
 
-    boss.start().catch(done)
+  it('clock monitoring error handling works', function (done) {
+    const config = { ...this.test.bossConfig, __test__throw_clock_monitoring: true }
+    const boss = new PgBoss(config)
+
+    boss.on('error', async () => {
+      boss.removeAllListeners()
+      await boss.stop()
+      done()
+    })
+
+    boss.start()
   })
 })