+++ test/configTest.js
@@ -35,34 +35,18 @@ describe('config', function () {
     await boss.stop()
   })
 
-  it('set pool config `poolSize`', async function () {
-    const poolSize = 4
-    const boss = await helper.start({ ...this.test.bossConfig, poolSize })
-
-    assert(boss.db.config.poolSize === poolSize)
-    assert(boss.db.pool.options.max === poolSize)
-
-    await boss.stop()
+  it('should not allow calling job instance functions if not started', async function () {
+    const boss = new PgBoss(this.test.bossConfig)
+    try {
+      await boss.publish('queue1')
+      assert(false)
+    } catch {}
   })
 
-  it('set pool config `max`: `poolSize` === `max`', async function () {
-    const max = 4
-    const boss = await helper.start({ ...this.test.bossConfig, max })
-
-    assert(boss.db.config.max === boss.db.config.poolSize)
-    assert(boss.db.config.max === max)
-    assert(boss.db.config.poolSize === max)
-    assert(boss.db.pool.options.max === max)
-
-    await boss.stop()
-  })
-
-  it('start() should fail if pgcrypto is not available', async function () {
+  it.skip('start() should fail if pgcrypto is not available', async function () {
     const database = 'pgboss_test1'
 
-    const db = await helper.getDb('postgres')
-
-    await db.executeSql(`CREATE DATABASE ${database}`)
+    await helper.createDb(database)
 
     const config = { ...this.test.bossConfig, database }
 
@@ -75,7 +59,6 @@ describe('config', function () {
       assert(err.message.includes('gen_random_uuid()'))
     }
 
-    await db.executeSql(`SELECT pg_terminate_backend( pid ) FROM pg_stat_activity WHERE pid <> pg_backend_pid() AND datname = '${database}'`)
-    await db.executeSql(`DROP DATABASE ${database}`)
+    await helper.tryDropDb(database)
   })
 })